updated websocket changes
